+++ joulescope_ui/widgets/__init__.py
@@ -25,6 +25,7 @@ from .hamburger import HamburgerWidget
 from .help import HelpWidget
 from .jls_info import JlsInfoWidget
 from .js220_cal import JS220CalibrationWidget
+from .js320_cal import JS320CalibrationWidget
 from .notes import NotesWidget
 from .progress_bar import ProgressBarWidget
 from .record_status import RecordStatusWidget