+++ CHANGELOG.md
@@ -26,9 +26,10 @@ This file contains the list of changes made to pyjoulescope_ui.
 2026 May 29
 
 * Added JS320 firmware update support.
+* Added JS320 offset calibration support.
 * Added support to release directly from GitHub Actions.
 * Bumped versions:
-  * pyjoulescope_driver from 2.1.0 to 2.1.1
+  * pyjoulescope_driver from 2.1.1 to 2.2.2
   * PySide6-QtAds from >=4.5.0.3 to >=4.5.0.5
   * PySide6 from >=6.11.0 to >=6.11.1